requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
openai>=1.3.0
tiktoken>=0.5.1
chromadb>=0.4.18
pydantic>=2.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
tenacity>=8.2.3
aiohttp>=3.8.6
pyyaml>=6.0.1
tqdm>=4.66.1
regex>=2023.10.3
nltk>=3.8.1
spacy>=3.7.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
jinja2>=3.1.2 